requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-Limiter
psycopg2-binary
python-dotenv
gunicorn